==> data-raw/03_lpi_index.R <==
# Script to calculare the Living Planet Index per taxa group and for all 
# taxa groups together

# load libraries
require(tidyr)
require(dplyr)
require(ggplot2)
require(EnvStats)

# load functions
source("data-raw/00_lpi_functions.R")

# import population growth rates
growthrates <- readRDS("data/lpi_growthrates.RDS") 
# remove entries before 1990 (why are these here if the prediction years in the previous script are only 1990 onwards???)
growthrates <- dplyr::filter(growthrates, year_pred >= 1990)


## get growth rates per species ----

# split into list by scientific name
growthrates_sp <- group_split(growthrates, scientific_name, year_pred)

# get mean per year per species
dt_gm_sp <- vector("list", length(growthrates_sp))
for(i in 1:length(dt_gm_sp)){
  if(nrow(growthrates_sp[[i]]) == 1){
    dt_gm_sp[[i]] <- growthrates_sp[[i]]
    } else{
    if(length(unique(growthrates_sp[[i]]$dt)) == 1){
      dt_gm_sp[[i]] <- growthrates_sp[[i]][1,]
    } else{
    # calculate geometric mean, per species
    temp <- dt_boot(10^growthrates_sp[[i]]$dt)
    dt_gm_sp[[i]] <- growthrates_sp[[i]][1,]
    dt_gm_sp[[i]]$dt <- log10(temp$gm)
    dt_gm_sp[[i]]$cilo <- log10(temp$cilo)
    dt_gm_sp[[i]]$cihi <- log10(temp$cihi)
  } 
  }
}

# bind rows back together
growthrates_sp <- bind_rows(dt_gm_sp)
# change Nan to NA
is.nan.data.frame <- function(x) do.call(cbind, lapply(x, is.nan))
growthrates_sp[is.nan(growthrates_sp)] <- NA


## LPI per taxa group ----

# split into taxa groups
growthrates_ls <- split(growthrates_sp, growthrates_sp$taxa)

# get mean per year per taxa
dt_gm_taxa <- vector("list", length(growthrates_ls))
for(i in c(1:length(dt_gm_taxa))){
  dt_gm_taxa[[i]] <- dt_gm_group(growthrates_ls[[i]], names(growthrates_ls)[i])
}

# calculate LPI per taxa with confidence intervals
for(i in 1:length(dt_gm_taxa)){
  dt_gm_taxa[[i]]$lpi <- calclpi(10^dt_gm_taxa[[i]]$gm)
  dt_gm_taxa[[i]]$lpi_cilo <- calclpi(10^dt_gm_taxa[[i]]$cilo)
  dt_gm_taxa[[i]]$lpi_cihi <- calclpi(10^dt_gm_taxa[[i]]$cihi)
}
ggplot(lpi_df, aes(x = year, y = lpi, col = taxa)) + geom_line()

## LPI for all VERTEBRATE groups ----

# get mean per year overall for vertebrates
dt_gm_tous <- dt_gm_group(filter(growthrates, taxa != "invertébrés"), 
                          groupname = "tous")
# calculate overall LPI with confidence intervals
dt_gm_tous$lpi <- calclpi(10^dt_gm_tous$gm)
dt_gm_tous$lpi_cilo <- calclpi(10^dt_gm_tous$cilo)
dt_gm_tous$lpi_cihi <- calclpi(10^dt_gm_tous$cihi)

# bind all groups together
lpi_df <- bind_rows(dt_gm_taxa, .id = NULL) %>%
  rbind(dt_gm_tous) %>%
  select(c(groupid, year, gm, cilo, cihi, lpi, lpi_cilo, lpi_cihi)) %>%
  rename(taxa = groupid)
ggplot(filter(lpi_df, taxa == "tous"), aes(x = year, y = lpi, col = taxa)) + geom_line()

## MANUALLY CHANGE FAKE DATA FOR TEST PURPOSES ---- to remove after the prototype ## -----------

# fix oiseaux and reptile uncertainty interval, which balloon outwards too rapidly
temp <- lpi_df[which(lpi_df$taxa == "oiseaux"), "lpi_cihi"] 
lpi_df[which(lpi_df$taxa == "oiseaux"), "lpi_cihi"]  <- lpi_df[which(lpi_df$taxa == "oiseaux"), "lpi"] + sort(runif(min = 0.1, max = 0.4, n = length(temp)))
lpi_df[which(lpi_df$taxa == "oiseaux"), "lpi_cilo"]  <- lpi_df[which(lpi_df$taxa == "oiseaux"), "lpi"] - sort(runif(min = 0.1, max = 0.4, n = length(temp)))
lpi_df[which(lpi_df$taxa == "reptiles"), "lpi_cihi"]  <- lpi_df[which(lpi_df$taxa == "reptiles"), "lpi"] + sort(runif(min = 0.1, max = 0.4, n = length(temp)))
lpi_df[which(lpi_df$taxa == "reptiles"), "lpi_cilo"]  <- lpi_df[which(lpi_df$taxa == "reptiles"), "lpi"] - sort(runif(min = 0.1, max = 0.4, n = length(temp)))

# save 
saveRDS(lpi_df, "data/lpi_index_taxa.RDS")


## LPI for each population ----

lpi_pops <- growthrates %>%
  group_by(org_event, scientific_name) %>% 
  # org_event is a unique ID per population in Living Planet Database
  # added scientific_name because the fake populations are all labelled "fake" in org_event
  mutate(lpi = calclpi(dt),
         lpi_cilo = calclpi(cilo),
         lpi_cihi = calclpi(cihi)) %>%
  mutate(trend = ifelse(round(dt, 4) < 1, "déclin", 
                        ifelse(round(dt, 4) > 1, "croissance", "stable"))) %>%
  ungroup()
saveRDS(lpi_pops, "data/lpi_index_populations.RDS")


## mean growth rate for each species ----

# split into list
pops_ls <- split(growthrates_sp, growthrates_sp$scientific_name)
pop_df <- data.frame("scientific_name" = names(pops_ls), 
                     "dt" = NA, cilo = NA, cihi = NA)
for(i in 1:length(pops_ls)){ # broken
  pop_df[i, 2:4] <- dt_boot(10^pops_ls[[i]]$dt) 
}

# add column to describe growth rate trend category 
trend_df <- pop_df %>%
  mutate(trend = ifelse(round(dt, 4) < 1, "déclin", 
                        ifelse(round(dt, 4) > 1, "croissance", "stable")))
temp <- growthrates[!duplicated(growthrates$scientific_name),]
trend_df <- inner_join(trend_df, 
                      select(temp, c(taxa, scientific_name, common_name)), 
                      by = "scientific_name")      
# copy table to create "tous" option
trend_tous <- trend_df
trend_tous$taxa <- "tous"
# combine with taxa-grouped trend table
trend_df <- rbind(trend_df, trend_tous)
saveRDS(trend_df, "data/lpi_trend_populations.RDS")
